vscode/requirements.txt
flytekit
flytekitplugins-flyin
flytekitplugins-envd